refactor: Use optional attributes and object type constraints
This should not break backwards compatibility nor cause problems but yet increase type and access safety of objects.

There are a few other places where this might come handy, right now it has been restricted to "lose" objects with a
finite set of configurations.

Possibly, it is also feasible to map and create a comprehensive list of all configurations for other objects.

Something to take into account is that this is only natively available on TF >= 1.3