feat: Adds Option to Ignore Changes to GSIs
This relates to this issue - https://github.com/hashicorp/terraform-provider-aws/issues/671.

When using autoscaling with a provisioned table that has a GSI
applying a TF change whilst the indices are scaled will reset
capacity, which can be dangerous. This change has an option
to ignore changes to global_secondary_index, which seems to be
the only way to deal with this issue at present.